requirements.txt
# Scraper core
zendriver>=0.15.3
beautifulsoup4>=4.12.3
tenacity>=8.3.0
python-dotenv>=1.0.1
requests>=2.32.3
Pillow>=11.0

# FastAPI microservice
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP client asíncrono (Webflow API)
httpx>=0.27.0

# Cloudinary staging host (intermediate upload so Webflow CDN keeps
# slug-based filenames when the token lacks assets scopes)
cloudinary>=1.40.0

# Scheduler
apscheduler>=3.10.4
sqlalchemy>=2.0.0

# Validación cron
croniter>=2.0.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
